Create Traffic_Light_Source_Code
The code declares the pin associated to which LED. It also times when the Arduino need to send voltage to the correct pin. The project has two street light. When one is on green or yellow, then the other is on red. Configuration are similar to an intersection street light configuration.